Added an Instance class
