+++ FlingEngine/Graphics/inc/LogicalDevice.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "FlingVulkan.h"
+
+namespace Fling
+{
+    class LogicalDevice
+    {
+
+    };
+
+}   // namespace Fling
