+++ FlingEngine/Graphics/inc/PhyscialDevice.h
@@ -1,0 +1,13 @@
+#pragma once
+
+#include "FlingVulkan.h"
+
+namespace Fling
+{
+
+    class PhysicalDevice
+    {
+        
+    };
+
+}   // namespace Fling
