+++ FlingEngine/Graphics/src/Instance.cpp
@@ -1,0 +1,203 @@
+#include "pch.h"
+#include "Instance.h"
+#include "FlingConfig.h"
+#include <GLFW/glfw3.h>
+
+namespace Fling
+{
+    Instance::Instance()
+    {
+        m_EnableValidationLayers = FlingConfig::GetBool("Vulkan", "EnableValidationLayers", false);
+		F_LOG_TRACE("[Renderer] m_EnableValidationLayers is {}", m_EnableValidationLayers);
+
+        CreateInstance();
+
+        SetupDebugMessages();
+        F_LOG_WARN("Vulkan Instance created!");
+    }
+
+    void Instance::CreateInstance()
+    {
+        if( m_EnableValidationLayers && !CheckValidationLayerSupport() )
+		{
+            F_LOG_FATAL( "Validation layers are requested, but not available!" );
+		}
+
+		// Basic app data that we can modify 
+		VkApplicationInfo appInfo = {};
+		// Most structs in Vulkan require you to specify the sType and 
+		// the pNext values 
+		appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
+		appInfo.pApplicationName = "Engine Bay 0.1";
+		appInfo.applicationVersion = VK_MAKE_VERSION(Version::EngineVersion.Major, Version::EngineVersion.Minor, Version::EngineVersion.Patch);
+		appInfo.pEngineName = "Fling Engine";
+		appInfo.engineVersion = VK_MAKE_VERSION(Version::EngineVersion.Major, Version::EngineVersion.Minor, Version::EngineVersion.Patch);
+		appInfo.apiVersion = VK_API_VERSION_1_0;
+
+		// Instance creation info, similar to how DX11 worked
+		VkInstanceCreateInfo createInfo = {};
+		createInfo.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
+		createInfo.pApplicationInfo = &appInfo;
+
+		// Because we are using GLFW as an extension interface for window creation
+		// Vulkan needs to know how many extensions are available
+		std::vector<const char*> extensions = GetRequiredExtensions();
+		createInfo.enabledExtensionCount = static_cast<uint32_t>( extensions.size() );
+		createInfo.ppEnabledExtensionNames = extensions.data();
+
+		// Determine global validation layer count
+		if( m_EnableValidationLayers )
+		{
+			createInfo.enabledLayerCount = static_cast<UINT32>( m_ValidationLayers.size() );
+			createInfo.ppEnabledLayerNames = m_ValidationLayers.data();
+		}
+		else 
+		{
+			createInfo.enabledLayerCount = 0;
+			createInfo.pNext = nullptr;
+		}
+
+		// Create the vulkan instance! Throw exception if it fails
+		if( vkCreateInstance( &createInfo, nullptr, &m_Instance ) != VK_SUCCESS )
+		{
+            F_LOG_FATAL( "Failed to create the Vulkan instance!" );
+		}
+
+		// Find out what extensions are available
+		uint32_t extensionCount = 0;
+		vkEnumerateInstanceExtensionProperties( nullptr, &extensionCount, nullptr );
+		std::vector<VkExtensionProperties> vk_extensions( extensionCount );
+		vkEnumerateInstanceExtensionProperties( nullptr, &extensionCount, vk_extensions.data() );
+
+		F_LOG_TRACE( "{} Extensions are supported!", extensionCount );
+		for( const auto& extension : vk_extensions )
+		{
+			F_LOG_TRACE( "\t{}", extension.extensionName );
+		}
+    }
+
+    bool Instance::CheckValidationLayerSupport()
+	{
+		// Get the list of available validation layers
+		// Validation layers are ways for us to choose what types of debugging
+		// features we want from the Vulkan SDK
+		UINT32 layerCount = 0;
+		vkEnumerateInstanceLayerProperties( &layerCount, nullptr );
+
+		std::vector<VkLayerProperties> availableLayers( layerCount );
+		vkEnumerateInstanceLayerProperties( &layerCount, availableLayers.data() );
+
+		for( const char* layerName : m_ValidationLayers )
+		{
+			bool layerFound = false;
+			for( const VkLayerProperties& layerProperties : availableLayers )
+			{
+				if( strcmp( layerName, layerProperties.layerName ) == 0 )
+				{
+					layerFound = true;
+					break;
+				}
+			}
+			if( !layerFound )
+			{
+				return false;
+			}
+		}
+
+		return true;
+	}
+
+    std::vector<const char*> Instance::GetRequiredExtensions()
+	{
+		UINT32 glfwExtensionCount = 0;
+		const char** glfwExtensions;
+		glfwExtensions = glfwGetRequiredInstanceExtensions( &glfwExtensionCount );
+
+		std::vector<const char*> extensions( glfwExtensions, glfwExtensions + glfwExtensionCount );
+
+		if( m_EnableValidationLayers ) 
+		{
+			extensions.push_back( VK_EXT_DEBUG_UTILS_EXTENSION_NAME );
+		}
+
+		return extensions;
+	}
+
+    void Instance::SetupDebugMessages()
+    {
+        if( !m_EnableValidationLayers ) { return; }
+
+		VkDebugUtilsMessengerCreateInfoEXT createInfo = {};
+		createInfo.sType = VK_STRUCTURE_TYPE_DEBUG_UTILS_MESSENGER_CREATE_INFO_EXT;
+		createInfo.messageSeverity = VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT;
+		createInfo.messageType = VK_DEBUG_UTILS_MESSAGE_TYPE_GENERAL_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_VALIDATION_BIT_EXT | VK_DEBUG_UTILS_MESSAGE_TYPE_PERFORMANCE_BIT_EXT;
+		createInfo.pfnUserCallback = DebugCallback;
+		createInfo.pUserData = nullptr; // Optional
+		
+		if( CreateDebugUtilsMessengerEXT( m_Instance, &createInfo, nullptr, &m_DebugMessenger ) != VK_SUCCESS )
+		{
+            F_LOG_FATAL( "Failed to set up debug messenger!" );
+		}
+    }
+
+    VKAPI_ATTR VkBool32 VKAPI_CALL Instance::DebugCallback(
+		VkDebugUtilsMessageSeverityFlagBitsEXT t_messageSeverity, 
+		VkDebugUtilsMessageTypeFlagsEXT t_messageType, 
+		const VkDebugUtilsMessengerCallbackDataEXT* t_CallbackData, 
+		void* t_UserData )
+	{
+		switch( t_messageSeverity )
+		{
+		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_WARNING_BIT_EXT:
+			F_LOG_WARN( "Validation layer: {}", t_CallbackData->pMessage );
+			break;
+		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_ERROR_BIT_EXT:
+		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_FLAG_BITS_MAX_ENUM_EXT:
+			F_LOG_ERROR( "Validation layer error: {}", t_CallbackData->pMessage );
+			break;
+		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_VERBOSE_BIT_EXT:
+		case VK_DEBUG_UTILS_MESSAGE_SEVERITY_INFO_BIT_EXT:
+		default:
+			F_LOG_TRACE( "Validation layer: {}", t_CallbackData->pMessage );
+			break;
+		}
+
+		return VK_FALSE;
+	}
+
+    VkResult Instance::CreateDebugUtilsMessengerEXT( VkInstance instance, const VkDebugUtilsMessengerCreateInfoEXT* pCreateInfo, const VkAllocationCallbacks* pAllocator, VkDebugUtilsMessengerEXT* pDebugMessenger )
+	{
+		PFN_vkCreateDebugUtilsMessengerEXT func = (PFN_vkCreateDebugUtilsMessengerEXT)vkGetInstanceProcAddr( instance, "vkCreateDebugUtilsMessengerEXT" );
+		if( func != nullptr ) 
+		{
+			return func( instance, pCreateInfo, pAllocator, pDebugMessenger );
+		}
+		else 
+		{
+			return VK_ERROR_EXTENSION_NOT_PRESENT;
+		}
+	}
+
+	void Instance::DestroyDebugUtilsMessengerEXT( VkInstance instance, VkDebugUtilsMessengerEXT debugMessenger, const VkAllocationCallbacks* pAllocator )
+	{
+		PFN_vkDestroyDebugUtilsMessengerEXT func = 
+			(PFN_vkDestroyDebugUtilsMessengerEXT)vkGetInstanceProcAddr( instance, "vkDestroyDebugUtilsMessengerEXT" );
+		if( func != nullptr ) 
+		{
+			func( instance, m_DebugMessenger, pAllocator );
+		}
+	}
+
+    Instance::~Instance()
+    {
+
+        // Debug messenger
+        if (m_EnableValidationLayers)
+        {
+            DestroyDebugUtilsMessengerEXT(m_Instance, m_DebugMessenger, nullptr);
+        }
+
+		vkDestroyInstance( m_Instance, nullptr );
+
+    }
+}   // namespace Fling
