+++ FlingEngine/Graphics/src/LogicalDevice.cpp
@@ -1,0 +1,7 @@
+#include "pch.h"
+#include "LogicalDevice.h"
+
+namespace Fling
+{
+
+}   // namespace Fling
